2
2

+++ src/main/java/hostel/reservation/ReservationController.java
@@ -9,13 +9,18 @@ import hostel.room.Room;
 import hostel.room.RoomService;
 import hostel.user.UserService;
 import hostel.web.models.User;
+import hostel.web.security.jwt.JwtUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 public class ReservationController {
@@ -28,7 +33,11 @@ public class ReservationController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private JwtUtils jwtUtils;
+
     @GetMapping(path = "/hotel/{hotelId}/room/{roomId}/reservation")
+    @PreAuthorize("hasRole('CLIENT') or hasRole('WORKER') or hasRole('ADMIN')")
     public ResponseEntity<List<ReservationDTO>> getReservations(@PathVariable long hotelId, @PathVariable long roomId) {
         Room room = roomService.findByHotelIdAndRoomId(hotelId, roomId)
                 .orElseThrow(() -> new HttpNotFoundException("Provided room not found"));
@@ -47,6 +56,7 @@ public class ReservationController {
     }
 
     @GetMapping(path = "/hotel/{hotelId}/room/{roomId}/reservation/{reservationId}")
+    @PreAuthorize("hasRole('CLIENT') or hasRole('WORKER') or hasRole('ADMIN')")
     public ResponseEntity<ReservationDTO> getReservation(
             @PathVariable long hotelId, @PathVariable long roomId, @PathVariable long reservationId) {
         Room room = roomService.findByHotelIdAndRoomId(hotelId, roomId)
@@ -57,11 +67,22 @@ public class ReservationController {
     }
 
     @PostMapping(path = "/hotel/{hotelId}/room/{roomId}/reservation")
-    public ResponseEntity<String> addReservation(@PathVariable long hotelId, @PathVariable long roomId,
+    @PreAuthorize("hasRole('CLIENT') or hasRole('WORKER') or hasRole('ADMIN')")
+    public ResponseEntity<String> addReservation(@RequestHeader("Authorization") String authorizationHeader, @PathVariable long hotelId, @PathVariable long roomId,
                                                  @RequestBody ReservationDTO reservationDTO) {
         if (!reservationDTO.notNull()) {
             throw new HttpUnprocessableContent();
         }
+        String token = "";
+        String[] authorizationHeaderArray = authorizationHeader.split(" ");
+        if (authorizationHeaderArray.length == 2 && "Bearer".equals(authorizationHeaderArray[0])) {
+            token = authorizationHeaderArray[1];
+        }
+        else
+        {
+            return ResponseEntity.badRequest().body("Authorization token wrong");
+        }
+        String userId = jwtUtils.getUserIdFromJwtToken(token);
 
         Room room = roomService.findByHotelIdAndRoomId(hotelId, roomId)
                 .orElseThrow(() -> new HttpBadRequestException("Provided room not found"));
@@ -77,7 +98,8 @@ public class ReservationController {
     }
 
     @PatchMapping(path = "/hotel/{hotelId}/room/{roomId}/reservation/{reservationId}")
-    public ResponseEntity<String> updateReservation(@PathVariable long hotelId, @PathVariable long roomId,
+    @PreAuthorize("hasRole('CLIENT') or hasRole('WORKER') or hasRole('ADMIN')")
+    public ResponseEntity<String> updateReservation(@RequestHeader("Authorization") String authorizationHeader, @PathVariable long hotelId, @PathVariable long roomId,
                                                     @PathVariable long reservationId,
                                                     @RequestBody ReservationDTO reservationDTO) {
         if (!reservationDTO.atLeastOneNotNull()) {
@@ -101,16 +123,33 @@ public class ReservationController {
     }
     
     @DeleteMapping(path = "/hotel/{hotelId}/room/{roomId}/reservation/{reservationId}")
-    public ResponseEntity<String> deleteReservation(@PathVariable long hotelId,
+    @PreAuthorize("hasRole('CLIENT') or hasRole('WORKER') or hasRole('ADMIN')")
+    public ResponseEntity<String> deleteReservation(@RequestHeader("Authorization") String authorizationHeader, @PathVariable long hotelId,
                                                     @PathVariable long roomId,
                                                     @PathVariable long reservationId) {
-        Room room = roomService.findByHotelIdAndRoomId(hotelId, roomId)
-                .orElseThrow(() -> new HttpNotFoundException("Provided room not found"));
-        hostel.web.models.User user = userService.findById(String.valueOf(1)).orElseThrow(() -> new HttpNotFoundException("Error with user"));
-        Reservation reservation = reservationService.findByRoomIdAndId(roomId, reservationId)
-                .orElseThrow(() -> new HttpNotFoundException("Provided reservation not found"));
-        
-        reservationService.delete(reservation);
-        return new ResponseEntity<>("Successfully deleted reservation", HttpStatus.OK);
+        String token = "";
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String[] authorizationHeaderArray = authorizationHeader.split(" ");
+        if (authorizationHeaderArray.length == 2 && "Bearer".equals(authorizationHeaderArray[0])) {
+            token = authorizationHeaderArray[1];
+        }
+        else
+        {
+            return ResponseEntity.badRequest().body("Authorization token wrong");
+        }
+        if ( authentication.getAuthorities().stream().anyMatch(role -> "ROLE_ADMIN".equals(role.getAuthority()) || "ROLE_WORKER".equals(role.getAuthority())))
+        {
+            String userId = jwtUtils.getUserIdFromJwtToken(token);
+            Room room = roomService.findByHotelIdAndRoomId(hotelId, roomId)
+                    .orElseThrow(() -> new HttpNotFoundException("Provided room not found"));
+            hostel.web.models.User user = userService.findById(String.valueOf(1)).orElseThrow(() -> new HttpNotFoundException("Error with user"));
+            Reservation reservation = reservationService.findByRoomIdAndId(roomId, reservationId)
+                    .orElseThrow(() -> new HttpNotFoundException("Provided reservation not found"));
+
+            reservationService.delete(reservation);
+            return new ResponseEntity<>("Successfully deleted reservation", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Access denied", HttpStatus.FORBIDDEN);
+
     }
 }
